+++ components/micro-gateway-core/src/main/ballerina/src/gateway/native/mutualSSL.bal
@@ -1,40 +1,47 @@
 import ballerinax/java;
 
-# Get the Alias name of the cert.
+# Get the Alias name of the certificate used in handshake.
 #
-# + cert - certicate
-# + trustStorePath - truststore location
-# + trustStorePassword - truststore password
-# + return - An error if an error occurs during the process or name of the certificate alias
+# + cert - Certicate used in Mutual SSL handshake
+# + return - Name of the certificate alias or error occurs during the process
 
-public function getAlias(string cert,string trustStorePath,string trustStorePassword) returns handle|error {
-    handle cert1 = java:fromString(cert);
-    handle trustStorePath1 = java:fromString(trustStorePath);
-    handle trustStorePassword1 = java:fromString(trustStorePassword);
-    handle|error certAlias = jgetAlias(cert1,trustStorePath1,trustStorePassword1);
+public function getAlias(string cert) returns handle|error {
+    handle certificate = java:fromString(cert);
+    handle|error certAlias = jgetAlias(certificate);
     return certAlias;
 }
 
-function jgetAlias(handle cert, handle trustStorePath,handle trustStorePassword) returns handle|error = @java:Method {
+function jgetAlias(handle cert) returns handle|error = @java:Method {
     name: "getAlias",
-    class: "org.wso2.micro.gateway.core.mutualssl.MutualsslRequestInvoker"
+    class: "org.wso2.micro.gateway.core.mutualssl.MutualsslWithoutLoadBalancerHeader"
 } external;
 
-# validate the certificate whether it is exist in the trustore.
+# Get the Alias name of the cert used in header append by load balancer.
 #
-# + cert - certicate
-# + trustStorePath - truststore location
-# + trustStorePassword - truststore password
-# + return - An error if an error occurs during the process or boolean true/false
-public function isExistCert(string cert,string trustStorePath,string trustStorePassword) returns boolean|error {
-    handle cert1 = java:fromString(cert);
-    handle trustStorePath1 = java:fromString(trustStorePath);
-    handle trustStorePassword1 = java:fromString(trustStorePassword);
-    boolean|error certt = jisExistCert(cert1,trustStorePath1,trustStorePassword1);
-    return certt;
+# + cert - Certicate append in the header
+# + return - Name of the certificate alias or error occurs during the process
+public function getAliasAFromHeaderCert(string cert) returns handle|error {
+    handle certificate = java:fromString(cert);
+    handle|error certAlias = jisExistCert(certificate);
+    return certAlias;
 }
 
-function jisExistCert(handle cert, handle trustStorePath,handle trustStorePassword) returns boolean|error = @java:Method {
-    name: "isExistCert",
-    class: "org.wso2.micro.gateway.core.mutualssl.MutualsslHeaderInvoker"
+function jisExistCert(handle cert) returns handle|error = @java:Method {
+    name: "getAliasAFromHeaderCert",
+    class: "org.wso2.micro.gateway.core.mutualssl.MutualsslWithLoadBalancerHeader"
+} external;
+
+# Load the trustore in keystore.
+#
+# + trustStorePath - truststore location
+# + trustStorePassword - truststore password
+function loadKeyStore(string trustStorePath,string trustStorePassword) {
+    handle trustStorePath1 = java:fromString(trustStorePath);
+    handle trustStorePassword1 = java:fromString(trustStorePassword);
+    jloadKeyStore(trustStorePath1, trustStorePassword1);
+}
+
+function jloadKeyStore(handle trustStorePath,handle trustStorePassword) = @java:Method {
+    name: "loadKeyStore",
+    class: "org.wso2.micro.gateway.core.mutualssl.LoadKeyStore"
 } external;
